Interfaces migrated to model
