--- frontend/src/app/interfaces/basic-info.ts
@@ -1,9 +1,0 @@
-/**
- * @description
- *    This interface holds the basic informtion view related fields.
- */
-export interface BasicInfo {
-  firstName: string;
-  lastName: string;
-  salutation: string;
-}
