--- frontend/src/app/interfaces/contact-info.ts
@@ -1,10 +1,0 @@
-/**
- * @description
- *    This interface holds the contact informtion view related fields.
- */
-export interface ContactInfo {
-  phoneNumber: string;
-  eMail: string;
-  linkedIn: string;
-  xing: string;
-}
