+++ frontend/src/app/model/forms-data.model.ts
@@ -1,9 +1,17 @@
 export interface FormsData {
   basicInfo: BasicInfo;
+  contactInfo: ContactInfo;
 }
 
 export interface BasicInfo {
   firstName: string;
   lastName: string;
   salutation: string;
+}
+
+export interface ContactInfo {
+  phoneNumber: string;
+  eMail: string;
+  linkedIn: string;
+  xing: string;
 }
